Rename ResizableNotebook to DatasetNotebook
