+++ frontend/src/metabase/query_builder/components/DatasetEditor/DatasetNotebook/index.js
@@ -1,0 +1,1 @@
+export { default } from "./DatasetNotebook";
