--- frontend/src/metabase/query_builder/components/DatasetEditor/ResizableNotebook/index.js
@@ -1,1 +1,0 @@
-export { default } from "./ResizableNotebook";
